[#30] Moving classes to separated modules
