--- src/main/java/br/com/ppm/commons/ToStringConstants.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2020 pedrotoliveira
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package br.com.ppm.commons;
-
-/**
- * Constants Used in ToStringBuilder Classes
- *
- * @author pedrotoliveira
- */
-public final class ToStringConstants {
-
-    public static final String OPEN_MAP_BRACKET = "Map{";
-    public static final String OPEN_BRACKET = "{";
-    public static final String CLOSE_BRACKET = "}";
-    public static final String OPEN_SQUARE_BRACKET = "[";
-    public static final String CLOSE_SQUARE_BRACKET = "]";
-    public static final String EQUAL = "=";
-    public static final String COMMA = ", ";
-    public static final boolean IGNORE_SUPER_TYPES = true;
-    public static final boolean NOT_IGNORE_SUPER_TYPES = false;
-
-    private ToStringConstants() {
-    }
-}
