--- src/test/java/br/com/ppm/commons/ToStringBuilderFactoryTest.java
@@ -1,26 +1,0 @@
-package br.com.ppm.commons;
-
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class ToStringBuilderFactoryTest {
-
-    @Before
-    public void setUp() throws Exception {
-    }
-
-    @After
-    public void tearDown() throws Exception {
-    }
-
-    @Test
-    public void of() {
-    }
-
-    @Test
-    public void create() {
-    }
-}
